refactor rename a few classes. Remove defaults from our own enums switchs.
