+++ lock/src/main/java/com/auth0/android/lock/enums/UsernameStyle.java
@@ -30,5 +30,5 @@ package com.auth0.android.lock.enums;
 public enum UsernameStyle {
     EMAIL,
     USERNAME,
-    USERNAME_OR_EMAIL
+    DEFAULT
 }
