+++ lock/src/test/java/com/auth0/android/lock/OptionsTest.java
@@ -153,7 +153,7 @@ public class OptionsTest {
     public void shouldUseDefaultUsernameStyle() {
         Options options = new Options();
         options.setAccount(auth0);
-        options.setUsernameStyle(UsernameStyle.USERNAME_OR_EMAIL);
+        options.setUsernameStyle(UsernameStyle.DEFAULT);
 
         Parcel parcel = Parcel.obtain();
         options.writeToParcel(parcel, 0);
